requirements.txt
async-timeout==5.0.1
attrs==23.2.0
autopep8==2.3.1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
cryptography==3.4.8
distro==1.7.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.12.4
flake8==7.1.1
future==0.18.3
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
mysqlclient==2.2.6
packaging==21.3
Pillow==9.0.1
platformdirs==3.10.0
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.3.0
pyparsing==3.2.0
pyperclip==1.9.0
pytest==7.4.4
python-dotenv==1.0.1
redis==5.0.8
rich==13.9.4
sentry-sdk==2.19.2
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==1.0.0
